meta-zaius: avsbus-control: Raise VDN rails max voltage
Testing showed that the VDN max voltage setting is capping the VCS rail
setpoints when those setpoints are written by the host over AVSbus.

Raise VRMs' upper limit for voltage on the VDN rails to a value that
does not interfere with VCS rail voltages.

This is a hacky workaround for what is possibly a logic error with the
VRMs' firmware and will be rolled back when the root cause is discovered
and corrected.

Signed-off-by: Xo Wang <xow@google.com>
Change-Id: I85fd94f77509bc51aa7909dd7f522163666040fb
Signed-off-by: Jason Luo <jason.cc.luo@mail.foxconn.com>